Clean up client side code Organize by pages and components Move commnly used components into seperate file Fix indentation issues Move css into sepeerate folder
