+++ src/client/components/mod.rs
@@ -1,0 +1,2 @@
+use super::*;
+pub mod nav_bar;
