+++ src/client/pages/css/splash.css
