+++ src/client/pages/mod.rs
@@ -1,0 +1,9 @@
+use super::*;
+
+pub mod getstarted;
+pub mod chat;
+pub mod test;
+pub mod manual;
+pub mod personality;
+pub mod privacypolicy;
+pub mod splash;
