+++ src/client/pages/privacypolicy.rs
@@ -1,10 +1,10 @@
 #![allow(non_snake_case)]
 
 use crate::client;
+use crate::client::components::nav_bar::top_bar;
+use crate::client::components::nav_bar::Navbar;
 
 use client::markdown_converter;
-use client::top_bar;
-use client::Navbar;
 use client::State;
 use dioxus::prelude::*;
 
@@ -12,7 +12,7 @@ use dioxus::prelude::*;
 pub fn Privacypolicy() -> Element {
     let state = use_context::<State>();
     let show_sidebar = state.scores().is_some();
-    let policy = include_str!("../../files/privacypolicy.md");
+    let policy = include_str!("../../../files/privacypolicy.md");
 
     rsx! {
         if show_sidebar {Navbar {active_chat: false}} else { { top_bar() } },
