+++ src/client/utils.rs
@@ -424,41 +424,3 @@ pub fn MessageList(mut msgs: MessageListProps) -> Element {
         }
     )
 }
-
-#[component]
-pub fn Navbar(active_chat: bool) -> Element {
-    rsx! {
-        nav {
-            position: "fixed",
-            z_index: "1000",
-
-            ul {
-                Link {
-
-                    to: Route::Home {},
-                    img {
-                        src: "logo.svg",
-                        alt: "Oceanchat Logo",
-                        class: "logo",
-                    }
-
-                    background_color: "transparent",
-                }
-                li {
-                    Link {
-                        to: Route::Chat {},
-                        "Chat",
-                        class: if active_chat { "active" } else { "" }
-                    }
-                }
-                li {
-                    Link {
-                        to: Route::Personality {},
-                        "My personality",
-                        class: if !active_chat { "active" } else { "" }
-                    }
-                }
-            }
-        }
-    }
-}
